+++ src/defaults.ts
@@ -32,7 +32,6 @@ export function hexagonalDefaults(
 ): RuleOptions {
   const defaults: RuleOptions = {
     rootDir: DEFAULTS.rootDir,
-    crossBoundaryStyle: DEFAULTS.crossBoundaryStyle,
     boundaries: [
       {
         identifier: '@domain',
@@ -106,7 +105,6 @@ export function simpleDefaults(
 ): RuleOptions {
   const defaults: RuleOptions = {
     rootDir: DEFAULTS.rootDir,
-    crossBoundaryStyle: DEFAULTS.crossBoundaryStyle,
     boundaries: [
       { identifier: '@domain', dir: 'domain', alias: '@domain' },
       { identifier: '@application', dir: 'application', alias: '@application' },